Summary: consolidate ut and ct common mocks Type: Cleanup Test Plan: Run CT and UT Jira: RIALTO-424
