+++ tests/common/mocks/GstWrapperFactoryMock.h
