+++ tests/common/mocks/GstWrapperMock.h
