+++ tests/common/mocks/WebAudioPlayerClientMock.h
@@ -2,7 +2,7 @@
  * If not stated otherwise in this file or this component's LICENSE file the
  * following copyright and licenses apply:
  *
- * Copyright 2022 Sky UK
+ * Copyright 2023 Sky UK
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,23 +17,24 @@
  * limitations under the License.
  */
 
-#ifndef FIREBOLT_RIALTO_COMMON_EVENT_THREAD_MOCK_H_
-#define FIREBOLT_RIALTO_COMMON_EVENT_THREAD_MOCK_H_
+#ifndef FIREBOLT_RIALTO_WEB_AUDIO_PLAYER_CLIENT_MOCK_H_
+#define FIREBOLT_RIALTO_WEB_AUDIO_PLAYER_CLIENT_MOCK_H_
 
-#include "IEventThread.h"
 #include <gmock/gmock.h>
+#include <memory>
 
-namespace firebolt::rialto::common
+#include "IWebAudioPlayerClient.h"
+#include "MediaCommon.h"
+
+namespace firebolt::rialto
 {
-class EventThreadMock : public IEventThread
+class WebAudioPlayerClientMock : public IWebAudioPlayerClient
 {
 public:
-    EventThreadMock() = default;
-    virtual ~EventThreadMock() = default;
-
-    MOCK_METHOD(void, flush, (), (override));
-    MOCK_METHOD(void, addImpl, (std::function<void()> && func), (override));
+    WebAudioPlayerClientMock() = default;
+    virtual ~WebAudioPlayerClientMock() = default;
+    MOCK_METHOD(void, notifyState, (WebAudioPlayerState state), (override));
 };
-} // namespace firebolt::rialto::common
+} // namespace firebolt::rialto
 
-#endif // FIREBOLT_RIALTO_COMMON_EVENT_THREAD_MOCK_H_
+#endif // FIREBOLT_RIALTO_WEB_AUDIO_PLAYER_CLIENT_MOCK_H_
