+++ tests/unittests/media/server/ipc/controlModuleService/ControlModuleServiceTestsFixture.h
@@ -20,10 +20,10 @@
 #ifndef CONTROL_MODULE_SERVICE_TESTS_FIXTURE_H_
 #define CONTROL_MODULE_SERVICE_TESTS_FIXTURE_H_
 
+#include "ClientMock.h"
 #include "ClosureMock.h"
 #include "ControlServiceMock.h"
 #include "IControlModuleService.h"
-#include "IpcClientMock.h"
 #include "IpcControllerMock.h"
 #include "IpcServerMock.h"
 #include "PlaybackServiceMock.h"
