+++ tests/unittests/media/server/ipc/mediaPipelineModuleService/MediaPipelineModuleServiceTestsFixture.h
@@ -20,9 +20,9 @@
 #ifndef MEDIA_PIPELINE_MODULE_SERVICE_TESTS_FIXTURE_H_
 #define MEDIA_PIPELINE_MODULE_SERVICE_TESTS_FIXTURE_H_
 
+#include "ClientMock.h"
 #include "ClosureMock.h"
 #include "IMediaPipelineModuleService.h"
-#include "IpcClientMock.h"
 #include "IpcControllerMock.h"
 #include "IpcServerMock.h"
 #include "MediaPipelineServiceMock.h"
