+++ tests/unittests/media/server/ipc/serverManagerModuleService/ServerManagerModuleServiceTestsFixture.h
@@ -20,8 +20,8 @@
 #ifndef RIALTO_SERVER_MANAGER_MODULE_SERVICE_TESTS_FIXTURE_H_
 #define RIALTO_SERVER_MANAGER_MODULE_SERVICE_TESTS_FIXTURE_H_
 
+#include "ClientMock.h"
 #include "ClosureMock.h"
-#include "IpcClientMock.h"
 #include "IpcControllerMock.h"
 #include "IpcServerMock.h"
 #include "RpcControllerMock.h"
